资讯页时间问题修改

+++ app/src/main/java/com/luuu/seven/adapter/ComicNewsListAdapter.kt
@@ -21,7 +21,7 @@ class ComicNewsListAdapter(layoutResId: Int, data: List<ComicNewsListBean>) :
         ifNotNull(helper, item, {helper, item ->
             helper.setText(R.id.tv_news_list_title, item.title)
             helper.setText(R.id.tv_news_list_authors, item.nickname)
-            helper.setText(R.id.tv_create_time, DateFormat.format("yyyy-MM-dd", (item.createTime * 1000).toLong()))
+            helper.setText(R.id.tv_create_time, DateFormat.format("yyyy-MM-dd", item.createTime*1000))
             helper.getView<ImageView>(R.id.iv_news_list_cover).loadImg(item.rowPicUrl)
         })
     }
